feat: code review 반영
- 추상 클래스 오버 엔지니어링 제거
- ConstraintsTranslator 인터페이스에 supports 메서드 추가
- QueryBuilder -> QueryTranslator
