+++ src/main/java/persistence/sql/ddl/common/StringConstants.java
@@ -1,9 +1,17 @@
 package persistence.sql.ddl.common;
 
 public class StringConstants {
+
     private StringConstants() {
         // Do Nothing
     }
 
     public static final String EMPTY_STRING = "";
+
+    public static final CharSequence SPACE = " ";
+
+    public static final String SCHEMA_TABLE_DELIMITER = ".";
+
+    public static final String COLUMN_DEFINITION_DELIMITER = ", ";
+
 }
