+++ src/main/java/persistence/sql/ddl/constraints/ConstraintsTranslator.java
@@ -1,0 +1,9 @@
+package persistence.sql.ddl.constraints;
+
+import java.lang.reflect.Field;
+
+public interface ConstraintsTranslator {
+    String getConstraintsFrom(Field field);
+
+    boolean supports(Field field);
+}
