--- src/main/java/persistence/sql/ddl/constraints/builder/ConstraintsBuilder.java
@@ -1,7 +1,0 @@
-package persistence.sql.ddl.constraints.builder;
-
-import java.lang.reflect.Field;
-
-public interface ConstraintsBuilder {
-    String getConstraintsFrom(Field field);
-}
