+++ src/main/java/persistence/sql/ddl/constraints/strategy/DefaultConstraintsStrategy.java
@@ -3,20 +3,24 @@ package persistence.sql.ddl.constraints.strategy;
 import java.lang.reflect.Field;
 import java.util.List;
 import java.util.stream.Collectors;
-import persistence.sql.ddl.constraints.builder.ConstraintsBuilder;
-import persistence.sql.ddl.constraints.builder.NotNullConstraintsBuilder;
-import persistence.sql.ddl.constraints.builder.UniqueConstraintsBuilder;
+import persistence.sql.ddl.common.StringConstants;
+import persistence.sql.ddl.constraints.impl.AutoIncrementConstraintsTranslator;
+import persistence.sql.ddl.constraints.ConstraintsTranslator;
+import persistence.sql.ddl.constraints.impl.NotNullConstraintsTranslator;
+import persistence.sql.ddl.constraints.impl.UniqueConstraintsTranslator;
 
 public class DefaultConstraintsStrategy implements ConstraintsStrategy {
-    private final List<ConstraintsBuilder> constraintsBuilders = List.of(
-        new UniqueConstraintsBuilder(),
-        new NotNullConstraintsBuilder()
+    private final List<ConstraintsTranslator> constraintsTranslators = List.of(
+        new AutoIncrementConstraintsTranslator(),
+        new UniqueConstraintsTranslator(),
+        new NotNullConstraintsTranslator()
     );
 
     @Override
     public String getConstraintsFrom(Field field) {
-        return constraintsBuilders.stream()
-            .map(constraintsBuilder -> constraintsBuilder.getConstraintsFrom(field))
-            .collect(Collectors.joining(" "));
+        return constraintsTranslators.stream()
+            .filter(constraintsTranslator -> constraintsTranslator.supports(field))
+            .map(constraintsTranslator -> constraintsTranslator.getConstraintsFrom(field))
+            .collect(Collectors.joining(StringConstants.SPACE));
     }
 }
