+++ src/test/java/persistence/sql/ddl/query/QueryTranslator1Test.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.impl;
+package persistence.sql.ddl.query;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,20 +9,19 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.slf4j.Logger;
 import persistence.sql.ddl.entity.Person1;
-import persistence.sql.ddl.QueryBuilder;
 
 @DisplayName("1단계 요구사항 - @Entity, @Id 어노테이션을 바탕으로 create 쿼리 만들어보기")
-class QueryBuilder1Test {
-    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryBuilder2Test.class);
+class QueryTranslator1Test {
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryTranslator2Test.class);
 
     private final Class<?> entityClass = Person1.class;
 
-    private final QueryBuilder queryBuilder = new DefaultQueryBuilder();
+    private final QueryTranslator queryTranslator = new QueryTranslator();
 
     @Test
     @DisplayName("@Entity, @Id 어노테이션을 바탕으로 create 쿼리 만들어보기")
     void createDDL() {
-        String ddl = queryBuilder.buildDDL(entityClass);
+        String ddl = queryTranslator.getCreateTableQuery(entityClass);
 
         log.debug("DDL: {}", ddl);
 
@@ -33,7 +32,7 @@ class QueryBuilder1Test {
     @Test
     @DisplayName("@Entity, @Id 어노테이션을 바탕으로 drop 쿼리 만들어보기")
     void buildDropQuery() {
-        String dropQuery = queryBuilder.buildDropQuery(entityClass);
+        String dropQuery = queryTranslator.getDropTableQuery(entityClass);
 
         log.debug("Drop query: {}", dropQuery);
 
@@ -43,7 +42,7 @@ class QueryBuilder1Test {
     @Test
     @DisplayName("클래스 정보를 바탕으로 테이블명 가져오기")
     void getTableNameByClassName() {
-        String tableName = queryBuilder.getTableNameFrom(entityClass);
+        String tableName = queryTranslator.getTableNameFrom(entityClass);
 
         log.debug("Table name: {}", tableName);
 
@@ -54,7 +53,7 @@ class QueryBuilder1Test {
     @Test
     @DisplayName("클래스 정보를 바탕으로 컬럼 선언문 가져오기")
     void getColumnDefinitionStatement() {
-        String columnDefinitionStatement = queryBuilder.getTableColumnDefinitionFrom(entityClass);
+        String columnDefinitionStatement = queryTranslator.getTableColumnDefinitionFrom(entityClass);
 
         log.debug("Column definition statement: {}", columnDefinitionStatement);
 
@@ -73,7 +72,7 @@ class QueryBuilder1Test {
     ) throws NoSuchFieldException {
         Field field = entityClass.getDeclaredField(fieldName);
 
-        String actualColumnDefinitionStatement = queryBuilder.getColumnDefinitionFrom(field);
+        String actualColumnDefinitionStatement = queryTranslator.getColumnDefinitionFrom(field);
 
         assertThat(actualColumnDefinitionStatement).isEqualTo(expectedColumnDefinitionStatement);
     }
