+++ src/test/java/persistence/sql/ddl/query/QueryTranslator2Test.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.impl;
+package persistence.sql.ddl.query;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,21 +9,20 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.slf4j.Logger;
 import persistence.sql.ddl.entity.Person2;
-import persistence.sql.ddl.QueryBuilder;
 
 @DisplayName("2단계 요구사항 - @Entity, @Id, @Column 어노테이션을 바탕으로 create 쿼리 만들어보기")
-class QueryBuilder2Test {
+class QueryTranslator2Test {
 
-    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryBuilder2Test.class);
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryTranslator2Test.class);
 
     private final Class<?> entityClass = Person2.class;
 
-    private final QueryBuilder queryBuilder = new DefaultQueryBuilder();
+    private final QueryTranslator queryTranslator = new QueryTranslator();
 
     @Test
     @DisplayName("@Entity, @Id, @Column 어노테이션을 바탕으로 create 쿼리 만들어보기")
     void createDDL() {
-        String ddl = queryBuilder.buildDDL(entityClass);
+        String ddl = queryTranslator.getCreateTableQuery(entityClass);
 
         log.debug("DDL: {}", ddl);
 
@@ -34,7 +33,7 @@ class QueryBuilder2Test {
     @Test
     @DisplayName("@Entity, @Id, @Column 어노테이션을 바탕으로 drop 쿼리 만들어보기")
     void buildDropQuery() {
-        String dropQuery = queryBuilder.buildDropQuery(entityClass);
+        String dropQuery = queryTranslator.getDropTableQuery(entityClass);
 
         log.debug("Drop query: {}", dropQuery);
 
@@ -44,7 +43,7 @@ class QueryBuilder2Test {
     @Test
     @DisplayName("클래스 정보를 바탕으로 테이블명 가져오기")
     void getTableNameByClassName() {
-        String tableName = queryBuilder.getTableNameFrom(entityClass);
+        String tableName = queryTranslator.getTableNameFrom(entityClass);
 
         log.debug("Table name: {}", tableName);
 
@@ -54,7 +53,7 @@ class QueryBuilder2Test {
     @Test
     @DisplayName("클래스 정보와 @Id, @Column 어노테이션을 바탕으로 컬럼 선언문 가져오기")
     void getColumnDefinitionStatement() {
-        String columnDefinitionStatement = queryBuilder.getTableColumnDefinitionFrom(entityClass);
+        String columnDefinitionStatement = queryTranslator.getTableColumnDefinitionFrom(entityClass);
 
         log.debug("Column definition statement: {}", columnDefinitionStatement);
 
@@ -74,7 +73,7 @@ class QueryBuilder2Test {
     ) throws NoSuchFieldException {
         Field field = entityClass.getDeclaredField(fieldName);
 
-        String actualColumnDefinitionStatement = queryBuilder.getColumnDefinitionFrom(field);
+        String actualColumnDefinitionStatement = queryTranslator.getColumnDefinitionFrom(field);
 
         assertThat(actualColumnDefinitionStatement).isEqualTo(expectedColumnDefinitionStatement);
     }
