+++ src/test/java/persistence/sql/ddl/query/QueryTranslator4Test.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.impl;
+package persistence.sql.ddl.query;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,20 +6,19 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import persistence.sql.ddl.entity.Person4;
-import persistence.sql.ddl.QueryBuilder;
 
 @DisplayName("4단계 요구사항 - @Entity, @Table(schema), @Id, @Column, @Transient 어노테이션을 바탕으로 drop 쿼리 만들어보기")
-class QueryBuilder4Test {
-    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryBuilder4Test.class);
+class QueryTranslator4Test {
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryTranslator4Test.class);
 
     private final Class<?> entityClass = Person4.class;
 
-    private final QueryBuilder queryBuilder = new DefaultQueryBuilder();
+    private final QueryTranslator queryTranslator = new QueryTranslator();
 
     @Test
     @DisplayName("@Entity, @Table(schema), @Id, @Column, @Transient 어노테이션을 바탕으로 drop 쿼리 만들어보기")
     void buildDropQuery() {
-        String dropQuery = queryBuilder.buildDropQuery(entityClass);
+        String dropQuery = queryTranslator.getDropTableQuery(entityClass);
 
         log.debug("Drop query: {}", dropQuery);
