+++ src/test/java/persistence/sql/ddl/query/QueryTranslatorTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.impl;
+package persistence.sql.ddl.query;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,22 +8,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.slf4j.Logger;
-import persistence.sql.ddl.entity.Person3;
-import persistence.sql.ddl.QueryBuilder;
+import persistence.sql.ddl.entity.Person;
+import persistence.sql.ddl.entity.Person4;
 
-@DisplayName("3단계 요구사항 - @Entity, @Table, @Id, @Column, @Transient 어노테이션을 바탕으로 create 쿼리 만들어보기")
-class QueryBuilder3Test {
+class QueryTranslatorTest {
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryTranslatorTest.class);
 
-    private static final Logger log = org.slf4j.LoggerFactory.getLogger(QueryBuilder3Test.class);
+    private final Class<?> entityClass = Person.class;
 
-    private final Class<?> entityClass = Person3.class;
-
-    private final QueryBuilder queryBuilder = new DefaultQueryBuilder();
+    private final QueryTranslator queryTranslator = new QueryTranslator();
 
     @Test
     @DisplayName("@Entity, @Table, @Id, @Column, @Transient 어노테이션을 바탕으로 create 쿼리 만들어보기")
     void createDDL() {
-        String ddl = queryBuilder.buildDDL(entityClass);
+        String ddl = queryTranslator.getCreateTableQuery(entityClass);
 
         log.debug("DDL: {}", ddl);
 
@@ -32,9 +30,20 @@ class QueryBuilder3Test {
     }
 
     @Test
+    @DisplayName("@Entity, @Table(schema), @Id, @Column, @Transient 어노테이션을 바탕으로 create 쿼리 만들어보기")
+    void createDDLWithSchema() {
+        String ddl = queryTranslator.getCreateTableQuery(Person4.class);
+
+        log.debug("DDL: {}", ddl);
+
+        assertThat(ddl)
+            .isEqualTo("CREATE TABLE schema.users (id BIGINT AUTO_INCREMENT, nick_name VARCHAR(255), old INTEGER, email VARCHAR(255) UNIQUE NOT NULL)");
+    }
+
+    @Test
     @DisplayName("@Entity, @Table, @Id, @Column, @Transient 어노테이션을 바탕으로 drop 쿼리 만들어보기")
     void buildDropQuery() {
-        String dropQuery = queryBuilder.buildDropQuery(entityClass);
+        String dropQuery = queryTranslator.getDropTableQuery(entityClass);
 
         log.debug("Drop query: {}", dropQuery);
 
@@ -42,9 +51,19 @@ class QueryBuilder3Test {
     }
 
     @Test
+    @DisplayName("@Entity, @Table(schema), @Id, @Column, @Transient 어노테이션을 바탕으로 drop 쿼리 만들어보기")
+    void buildDropQueryWithSchema() {
+        String dropQuery = queryTranslator.getDropTableQuery(Person4.class);
+
+        log.debug("Drop query: {}", dropQuery);
+
+        assertThat(dropQuery).isEqualTo("DROP TABLE schema.users");
+    }
+
+    @Test
     @DisplayName("클래스 정보와 @Table 어노테이션을 바탕으로 테이블명 가져오기")
-    void getTableNameByClassName() {
-        String tableName = queryBuilder.getTableNameFrom(entityClass);
+    void getTableName() {
+        String tableName = queryTranslator.getTableNameFrom(entityClass);
 
         log.debug("Table name: {}", tableName);
 
@@ -52,9 +71,19 @@ class QueryBuilder3Test {
     }
 
     @Test
+    @DisplayName("클래스 정보와 @Table(schema) 어노테이션을 바탕으로 테이블명 가져오기")
+    void getTableNameWithSchema() {
+        String tableName = queryTranslator.getTableNameFrom(Person4.class);
+
+        log.debug("Table name: {}", tableName);
+
+        assertThat(tableName).isEqualTo("schema.users");
+    }
+
+    @Test
     @DisplayName("클래스 정보와 @Id, @Column, @Transient 어노테이션을 바탕으로 컬럼 선언문 가져오기")
     void getColumnDefinitionStatement() {
-        String columnDefinitionStatement = queryBuilder.getTableColumnDefinitionFrom(entityClass);
+        String columnDefinitionStatement = queryTranslator.getTableColumnDefinitionFrom(entityClass);
 
         log.debug("Column definition statement: {}", columnDefinitionStatement);
 
@@ -74,7 +103,7 @@ class QueryBuilder3Test {
     ) throws NoSuchFieldException {
         Field field = entityClass.getDeclaredField(fieldName);
 
-        String actualColumnDefinitionStatement = queryBuilder.getColumnDefinitionFrom(field);
+        String actualColumnDefinitionStatement = queryTranslator.getColumnDefinitionFrom(field);
 
         assertThat(actualColumnDefinitionStatement).isEqualTo(expectedColumnDefinitionStatement);
     }
